==> CODE/01_webscrape_GNP_fill_data.ipynb <==
Zach Tretter

June 2020

--------


# Step 1 - Generate the Data

Workbook Scope : Scrape the historical data for GLacier National Park from the website.  Use requests.post because the URL does not change when selecting a given month/year combination.  The output is will be a CSV file.  There are 13 campgrounds, 20 years, and 5 months available.

import requests
from bs4 import BeautifulSoup

import time
import datetime as dt
import calendar

import os
import pandas as pd
import numpy as np

## Webscrape Parent Page for Campground Status
1. Create a Beautiful Soup for the URL
2. Extract tags with campground names
3. Build the list of names and list of links

#### Instantiate Soup Object

# URL for Glacier National Park Campground Status
base_url = 'https://www.nps.gov/applications/glac/cgstatus/cgstatus.cfm'

# Create a request to the URL
res = requests.get(base_url)

# Create a beautiful soup object
soup = BeautifulSoup(res.content, 'lxml')

#### Extract Tags of Interest

# The third table contains the list of campgrounds
table = soup.find_all('table')[3]

# All campground names are in href tags
campgrounds = table.find_all('a')

#### Build Lists of Names and Links

# Build a list of campground names
names = []

# Build a list of links to the campground status
status_links = []

# Iterate through the campgrounds to build these lists
for i in campgrounds:
    names.append(i.getText())
    status_links.append(i.get('href'))    

#### Verify Output

# Verify the names and status links
for i in range(len(names)):
    print(names[i])
    print(status_links[i])
    print("\n")

## Webscrape Each Campground

http://jonathansoma.com/lede/foundations/classes/friday%20sessions/advanced-scraping-form-submissions-completed/

#### Build List of Full Status Link

# The base url for each campground's status page
status_url = 'https://www.nps.gov/applications/glac/cgstatus/'

# Create the list of links to status
fulllink_campgroundstatus_byname = []

for i in status_links:
    fulllink_campgroundstatus_byname.append(status_url + i)

# View Links
fulllink_campgroundstatus_byname

#### Identify range of parameters for requests.post

# Parameters for requests.post
available_years = [str(i) for i in np.arange(2000,2020)]
available_months = [str(i) for i in np.arange(5,10)]

total_calls = len(names) * len(available_years) * len(available_months)
print(f'Total requests to NPS website will be {total_calls}')
print(f'Using a time.sleep of 0.25 seconds this will require at least \n {round(total_calls/4 / 60,1)} minutes to run')

#### Define Function to Remove Escape Characters

def strip_escape_chars(text):
    '''
    getText() on an element in the 'Current and Historic Campground Fill Times' looks like this:
    '\n3\n\r\n\t\t\t\t\t\t\t11:00am\r\n\t\t\t\t\t\t\t\n\n'
    This function cleans it to: '3 11:00am'
    A date with no fill up time will just be a number (the day of the month)
    '''
    return text.strip().replace("\n","").replace('\t',"").replace('\r'," ")

#### Define Function to Extract Historic Fill Times

def retrieve_historic(link_to_campground,
                     months = available_months,
                     years = available_years):
    
    start_time = time.time()
    
    fill_times = []
    
    # Grab the Campground Name
    req = requests.get(link_to_campground)   
    soup = BeautifulSoup(req.text,'html.parser')
    title = soup.find('title').getText()
    name = title.split(" - ")[1].replace(" Campground Information","")

    # For each possible year, 2000 to 2019
    for specific_year in years:
        
        print(f'Examine {name} in {specific_year}')
        
        # For each possible month, May to September
        for specific_month in months:  
            
            # Pause between each loop to not trigger an adverse response from NPS
            time.sleep(0.5)
        
            # Declare the parameters for this particular request
            params = {'selectmm': specific_month,
                      'selectyy': specific_year}
            
            # Create the POST request
            response = requests.post(link_to_campground, params)
            
            # Create the Soup output
            soup = BeautifulSoup(response.text,'html.parser')
            
            # The third table contains the data of interest
            content = soup.find_all('table')[2]
            
            # Month-Year is the first cell
            month_text, year_text = strip_escape_chars(content.find_all('td')[0].getText()).split(" - ")
            month_num = int(dt.datetime.strptime(month_text, "%B").month)
            year = int(year_text)
            
            # Entries in the table start on the 9th cell
            # The first is the header "month - year"
            # The next seven are the days of the week

            for index, i in enumerate(content.find_all('td')[8:]):
                clean = strip_escape_chars(i.getText()).split(" ")

                # Ignore empty cells in the first row (e.g Wed when the month starts on a Fri)
                if clean == ['']:
                    # Break if we've gone through our dates
                    if index > 30:
                        break
                    else:
                        pass

                # For valid table entries, identify the fill up time or False if not
                else:
                    day = int(clean[0])
                    date = dt.date(year, month_num, day)
                    
                    if len(clean)==1:
                        did_fill = 0
                        fill_time = None
                        
                    else:
                        did_fill = 1
                        fill_time = clean[1]
#                     print(name, year,month,date,fill)
                    
                    entry = {
                        'cg_name': name,
                        'date': date,
                        'did_fill': did_fill,
                        'fill_time': fill_time,
                    }

                    fill_times.append(entry)
            
    print(f'Elapsed time {round(time.time() - start_time,0)} secs') 
                              
    return pd.DataFrame(fill_times)

### Build the Dataframe

df = pd.DataFrame()

for campground in fulllink_campgroundstatus_byname:
    function_output = retrieve_historic(campground)
    df = df.append(function_output)

df

### Export to CSV

df.to_csv('../01_filltimes_raw.csv')
