==> CODE/02_clean_target_data.ipynb <==
Zach Tretter

June 2020

--------


# Step 2 - Target Data Clean Up

Workbook Scope : Process the CSV that has the historic fill time data.  Specifically add additional informative date columns (e.g. year, month, day of the year, etc) and also identify invalid dates (false negatives).

----------

#### Contents
* [Make Campground Nicknames](#Campground-Nicknames)
* [Clean up Fill Time Column](#Clean-up-Fill-Time-Column)
* [Create Additional Date Columns](#Create-Additional-Date-Columns)
* [Identify Logan Pass Open/Close Dates](#Logan-Pass-Dates)
* [Identify Invalid Dates](#Identify-Invalid-Dates)
* [Identify Wildfire Closures](#Wildfire-Dates)
* [Look for Anomalies](#Look-for-Anomalies)
* [Export to CSV](#Export-to-CSV)

import requests
from bs4 import BeautifulSoup

import time
from datetime import datetime as dt
from datetime import timedelta
import calendar

import matplotlib.pyplot as plt

import os
import pandas as pd
import numpy as np

#### Import the raw csv

df = pd.read_csv('../data/01_filltimes_raw.csv')
df = df.drop(columns='Unnamed: 0')

# View the dataframe
df.info()

df['available']=1

## Campground Nicknames

#### Shorter Names for Campgrounds

# Shorter single-word names for each campground

nicknames = {
    'Apgar': 'Apga',
    'Avalanche': 'Aval',
    'Bowman Lake': 'BoLa',
    'Cut Bank': 'CuBa',
    'Fish Creek': 'FiCr',
    'Kintla Lake': 'KiLa',
    'Logging Creek': 'LoCr',
    'Many Glacier': 'MaGl',
    'Quartz Creek': 'QuCr',
    'Rising Sun': 'RiSu',
    'Sprague Creek': 'SpCr',
    'St. Mary': 'StMa',
    'Two Medicine': 'TwoM'
}

df['nickname'] = df['cg_name'].map(nicknames)

df['key_name_date'] = df['date']+ '_' + df['nickname']

## Clean up Fill Time Column

#### Create a column for time as in 24 hour format

# The function to do it

def time_as_24hr_string(fill_time):
    if type(fill_time) == str:
        return dt.strftime(dt.strptime(fill_time, '%I:%M%p'), "%H:%M")
    else:
        return 0

# Create this column in the dataframe

df['time_24'] = df['fill_time'].apply(time_as_24hr_string)

#### Create a column for time as numerical hours since midnight

# The function to do it

def time_as_hours_from_midnight(fill_time):
    time_24 = time_as_24hr_string(fill_time)
    
    if time_24 != 0:
        hour_val, minute_val = time_24.split(":")
        secs_elapsed = timedelta(hours = int(hour_val), minutes = int(minute_val)).seconds
        return round(secs_elapsed / 3600,2)
    
    else:
        return int(0)

# Create this column in the dataframe

df['hours_after_midnight'] = df['fill_time'].apply(time_as_hours_from_midnight)

## Create Additional Date Columns

#### Make the 'date' column of type date

df['date'] = pd.to_datetime(df['date'])

#### Isolate specific date attributes to include

https://strftime.org/

* Year (e.g. '2017')
* Month as Number (e.g. '09')
* Month as Text (e.g. 'Sep')
* Day of Year (e.g. '173')
* Day of Month (e.g. '4')
* Weekday (e.g. 'Sat')

# Isolate Year
df['year'] = df['date'].apply(
    lambda x: dt.strftime(x, "%Y")).astype(int)

# Isolate Month as Number
df['month_num'] = df['date'].apply(
    lambda x: dt.strftime(x, "%m")).astype(int)

# Isolate Month as Text
df['month_text'] = df['date'].apply(
    lambda x: dt.strftime(x, "%b"))

# Isolate Day of Year
df['day_of_year'] = df['date'].apply(
    lambda x: dt.strftime(x, "%j")).astype(int)

# Isolate Day of Month
df['day_of_month'] = df['date'].apply(
    lambda x: dt.strftime(x, "%d")).astype(int)

# Isolate Day of Week as Text
df['weekday'] = df['date'].apply(
    lambda x: dt.strftime(x, "%a"))

# Key for Year_Month combo
df['key_year_month'] = df['year'].astype(str) + "_" + df['month_text']

#### Identify if High Season (July and August)

df['high_season']=((df['month_num']==7) | (df['month_num']==8)).astype(int)

#### Identify if Fri/Sat Night

df['weekend'] = ((df['weekday']=='Fri') | (df['weekday']=='Sat')).astype(int)

#### Identify if Date is day before, after, or of a Holiday

df['holiday']=0

# Independence Day
df.loc[(df['month_num']==7) & (
    (df['day_of_month']==3) |
    (df['day_of_month']==4) |
    (df['day_of_month']==5)),
      'holiday'] = 1

# Labor day is the first monday in September
labor_days = df.loc[(df['month_num']==9) 
                    & (df['day_of_month']<8) 
                    & (df['weekday']=='Mon')
                   ]['date'].unique()

# Memorial day is the last monday in May
memorial_days = df.loc[(df['month_num']==5) 
                       & (df['day_of_month'] > 24) 
                       & (df['weekday']=='Mon')
                      ]['date'].unique()

holiday_dates = []

for i in memorial_days:
    holiday_dates.append(i)
    holiday_dates.append(i - np.timedelta64(1,'D'))
    holiday_dates.append(i + np.timedelta64(1,'D'))
    
for i in labor_days:
    holiday_dates.append(i)
    holiday_dates.append(i - np.timedelta64(1,'D'))
    holiday_dates.append(i + np.timedelta64(1,'D'))

# Convert holiday to 1 for these dates
for day in holiday_dates:
    df.loc[df['date']==day,'holiday']==1

## Logan Pass Dates

#### Identify when Going to the Sun Road is Open

# Set every row as zero
df['logan_pass'] = 0

# Built in excel because its only 20 dates
# For that year, logan pass is open between these dates, as represent by a day of the year
logan_pass_dates = {
    2000:[148,290],
    2001:[158,295],
    2002:[179,301],
    2003:[150,301],
    2004:[151,292],
    2005:[142,304],
    2006:[174,296],
    2007:[182,259],
    2008:[184,294],
    2009:[177,291],
    2010:[175,291],
    2011:[194,261],
    2012:[171,289],
    2013:[172,266],
    2014:[183,265],
    2015:[170,291],
    2016:[168,286],
    2017:[179,274],
    2018:[174,274],
    2019:[174,274]
}

for YEAR in logan_pass_dates.keys():
    OPENDATE = logan_pass_dates[YEAR][0]
    CLOSEDATE = logan_pass_dates[YEAR][1]

    df.loc[(df['year'] == YEAR) & 
           ( (df['day_of_year'] >= OPENDATE)
            & (df['day_of_year'] <= CLOSEDATE) )
           , 'logan_pass'] = 1

## Identify Invalid Dates

# Examine a specific campground in detail

# specific_campground = 'Two Medicine'
# df.loc[(df['did_fill']==1) & 
#        (df['cg_name']==specific_campground)].sort_values('day_of_year')[['date','day_of_month']]

#### Earliest and Latest Date a Campground Ever Filled

max_min_fill = pd.DataFrame()

for campground in df['cg_name'].unique():
    max_min_fill = max_min_fill.append(
        df.loc[
            # Look at dates when the campground filled
            (df['did_fill']==1) & 
            
            # Look at a specific campground
            (df['cg_name']==campground),
            
            # Return these columns
            ['cg_name','date','day_of_year','did_fill','fill_time','month_text','day_of_month']
            
            # Sort by the the day of the year to get the earliest first and the latest last
            # .iloc returns the first and last element
        ].sort_values('day_of_year').iloc[[0,-1]]
    )

max_min_fill

#### Built list of Open/Close Dates within Excel

cg_open_close = [
    ['Apgar',1,366],
    ['Avalanche',168,262],
    ['Bowman Lake',138,258],
    ['Cut Bank',148,269],
    ['Fish Creek',152,248],
    ['Kintla Lake',148,268],
    ['Logging Creek',180,268],
    ['Many Glacier',139,269],
    ['Quartz Creek',180,366],
    ['Rising Sun',154,256],
    ['Sprague Creek',126,266],
    ['St. Mary',1,366],
    ['Two Medicine',146,269]
]

# Set to 1
df['available'] = 1

#### Update the Dataframe

for entry in cg_open_close:
    
    CAMPGROUND = entry[0]
    OPENDATE = entry[1]
    CLOSEDATE = entry[2]

    df.loc[
        (df['cg_name'] == CAMPGROUND) &
        ( (df['day_of_year'] < OPENDATE) 
         | (df['day_of_year'] > CLOSEDATE) ),
         'available'] = 0

#### Verify no Fill Dates are Classified as Unvailable

pd.crosstab(df['available'],
            df['did_fill'])

## Wildfire Dates


#### Create Column where are all dates are zero

df['fire_closure'] = 0

#### 2018 Howe Ridge Fire

howe_ridge_fire =[
['Apgar',226,365],
['Avalanche',225,365],
['Bowman Lake',0,0],
['Cut Bank',0,0],
['Fish Creek',225,365],
['Kintla Lake',0,0],
['Logging Creek',0,0],
['Many Glacier',0,0],
['Quartz Creek',0,0],
['Rising Sun',0,0],
['Sprague Creek',225,365],
['St. Mary',0,0],
['Two Medicine',0,0],
]

for entry in howe_ridge_fire:
    
    CAMPGROUND = entry[0]
    START = entry[1]
    FINISH = entry[2]

    df.loc[
        ( (df['cg_name'] == CAMPGROUND) & (df['year']==2018) ) &
        ( (df['day_of_year'] >= START) 
         & (df['day_of_year'] <= FINISH) ),
         'fire_closure'] = 1

#### 2017 Sprague Fire

sprague_fire =[
['Apgar',246,365],
['Avalanche',246,365],
['Bowman Lake',0,0],
['Cut Bank',0,0],
['Fish Creek',246,365],
['Kintla Lake',0,0],
['Logging Creek',0,0],
['Many Glacier',0,0],
['Quartz Creek',0,0],
['Rising Sun',0,0],
['Sprague Creek',246,365],
['St. Mary',0,0],
['Two Medicine',0,0]

]

for entry in sprague_fire:
    
    CAMPGROUND = entry[0]
    START = entry[1]
    FINISH = entry[2]

    df.loc[
        ( (df['cg_name'] == CAMPGROUND) & (df['year']==201) ) &
        ( (df['day_of_year'] >= START) 
         | (df['day_of_year'] <= FINISH) ),
         'fire_closure'] = 1

#### 2015 Reynolds Fire

reynolds_fire=[
['Apgar',0,0],
['Avalanche',0,0],
['Bowman Lake',0,0],
['Cut Bank',0,0],
['Fish Creek',0,0],
['Kintla Lake',0,0],
['Logging Creek',0,0],
['Many Glacier',0,0],
['Quartz Creek',0,0],
['Rising Sun',203,365],
['Sprague Creek',0,0],
['St. Mary',203,211],
['Two Medicine',0,0],
]

for entry in reynolds_fire:
    
    CAMPGROUND = entry[0]
    START = entry[1]
    FINISH = entry[2]

    df.loc[
        ( (df['cg_name'] == CAMPGROUND) & (df['year']==2015) ) &
        ( (df['day_of_year'] >= START) 
         & (df['day_of_year'] < FINISH) ),
         'fire_closure'] = 1

### Reset Available to zero on these fire closure dates

df.loc[df['fire_closure']==1,'available'] = 0

#### Verify by Crosstab

pd.crosstab(df['fire_closure'],
            df['did_fill'])

## Look for Anomalies

df[df['hours_after_midnight']>0].sort_values('hours_after_midnight')

#### Manual Corrections

df.loc[(df['key_name_date']=='2018-07-28_QuCr'),'hours_after_midnight'] = 12.82

## Export to CSV

df.drop(columns = ['fire_closure'],inplace= True)

df.to_csv('../data/02_filltimes_clean.csv')
