==> CODE/04_integrate_dataframes.ipynb <==
Zach Tretter

June 2020

# Step 4 - Build Overall Dataframe

### Read in Individual Dataframes
* [Fill Times](#Read-in-Fill-Times)
* [Weather](#Read-in-Weather-Data)
* [Visits](#Read-in-Visits-Data)
* [Air Quality](#Read-in-Air-Quality-Data)
* [Campground Attributes](#Read-in-Campground-Attributes)

### Integrate Dataframes

* Fill Times as foundation of [Final Dataframe](#Instantiate-final-df-as-df_fill)
* Left Join Dataframe with Air Quality on Date (AQI is same for all campsites)
* Left Join Dataframe with Monthly Visits on 'key_year_month') (e.g. YYYY_MMM)
* Left Join Dataframe with Campground Attributes on 'cg_name' (Campground Name)
    * Can now create 'supply' variable
    * Can now create 'date_wxstation'
* Left Join Dataframe with Weather on 'date_wxstation'
* Drop columns no longer needed

### Export to CSV
* [Export](#Export-Dataframe)

import pandas as pd
import os
import numpy as np

-------------

## Read in Fill Times

# Read in the data
df_fill = pd.read_csv('../data/02_filltimes_clean.csv')

df_fill.drop(columns='Unnamed: 0',inplace=True)

# Convert date to a datetype
df_fill['date'] = pd.to_datetime(df_fill['date'])

df_fill = df_fill.set_index('key_name_date');

#### View Dataframe

df_fill.T

-------------

## Read in Weather Data

df_weather = pd.read_csv('../data/03a_weather_clean.csv')
df_weather.drop(columns=["Unnamed: 0"],inplace=True)

df_weather = df_weather.set_index('date_wxstation');

#### Drop Irrelevant Columns

df_weather.drop(columns = [
    'LATITUDE',
    'LONGITUDE',
    'wx_station'
],inplace=True)

#### View Dataframe

df_weather.T

-------------

## Read in Visits Data

df_visits = pd.read_csv('../data/03b_monthly_visits_clean.csv')
df_visits.set_index('key_year_month',inplace=True)
df_visits = pd.DataFrame(df_visits['visits'].astype(int))

#### View Dataframe

df_visits

----------

## Read in Air Quality Data

df_aqi = pd.read_csv('../data/03c_air_quality_clean.csv')

#### Convert date to datetime and set date as index

df_aqi['date'] = pd.to_datetime(df_aqi['date'])
df_aqi.set_index('date',inplace=True)
df_aqi.sort_index(inplace=True)

#### Drop Irrelevant Columns

df_aqi.drop(columns=['Main Pollutant','date_aqi'],inplace=True)

#### View Dataframe

df_aqi

## Read in Campground Attributes

df_cg_attributes = pd.read_csv('../data/03d_campground_attributes_clean.csv',
                              index_col = 'cg_name')

#### View Dataframe

df_cg_attributes

--------


# Integrate Dataframes
[Back-to-top](#Back-to-Top)

#### Instantiate final df as df_fill

df = df_fill.copy()

df = df.sort_values(['date',
                     'cg_name'])

#### Join with AQI on 'date'

df = pd.merge(df,
              df_aqi,
              left_on = 'date',
              right_index = True,
              how = 'left')

#### Join with Monthly Visits on 'key_year_month'

df = pd.merge(df,
              df_visits,
              left_on = 'key_year_month',
              right_index = True,
              how = 'left')

#### Join with Campground Attributes at 'cg_name'

df = pd.merge(df,
              df_cg_attributes,
              left_on = 'cg_name',
              right_index = True,
              how = 'left')

#### Create 'Supply' Variable and Integrate

df['cg_supply'] = df['sites'] * df['available']
df_supply = pd.DataFrame(df.groupby('date')['cg_supply'].sum())

df = pd.merge(df,
             df_supply,
             left_on = 'date',
             right_index = True,
             how = 'left')

# I don't get how this works but it does?
df['cg_supply_y'] = df['cg_supply_y'] - df['sites'] * df['available']

#### Create 'date_wxstation' column to enable integration with weather dataframe

df['date_wxstation'] = df['date'].astype(str) + '_' + df['nearest_wx_station']

#### Join with Weather on 'date_wxstation'

df = pd.merge(df,
             df_weather,
             left_on = 'date_wxstation',
             right_index = True,
             how = 'left')

#### Drop Columns not of Interest

df = df.drop(columns = [
    
    # Fill Dates
    'fill_time',
    'nickname',
    'key_year_month',
    
    
    # Campground Attributes
    'nearest_wx_station',
    'cg_supply_x',
    
    # From Weather Data
    'date_wxstation',
    'date_y',
    
])

#### Examine Dataframe

df.T

# Export Dataframe

[Back-to-top](#Back-to-Top)

df.to_csv('../data/04_Full_DataFrame.csv')
